requirements.txt
slack-bolt>=1.18.0
python-dotenv>=1.0.0
certifi>=2023.0.0